Move error images to assets directory and convert to WebP
